Consilidate pcr and sha512 functions
